--- packages/neos-ui/src/Containers/EditModePanel/Panel/slick-styles.vanilla-css
@@ -1,158 +1,0 @@
-/* Slick Slider */
-.slick-slider
-{
-    position: relative;
-    display: block;
-    box-sizing: border-box;
-    user-select: none;
-    -webkit-touch-callout: none;
-    touch-action: pan-y;
-    -webkit-tap-highlight-color: transparent;
-}
-.slick-list
-{
-    position: relative;
-    display: block;
-    overflow: hidden;
-    margin: 0;
-    padding: 0;
-}
-.slick-list:focus
-{
-    outline: none;
-}
-.slick-list.dragging
-{
-    cursor: pointer;
-    cursor: hand;
-}
-
-.slick-slider .slick-track,
-.slick-slider .slick-list
-{
-    transform: translate3d(0, 0, 0);
-}
-.slick-track
-{
-    position: relative;
-    top: 0;
-    left: 0;
-    display: block;
-    width: 5000px; /* Set initial width because Slick does not set it correct. (See https://github.com/neos/neos-ui/pull/1591) */
-}
-.slick-track:before,
-.slick-track:after
-{
-    display: table;
-
-    content: '';
-}
-.slick-track:after
-{
-    clear: both;
-}
-.slick-loading .slick-track
-{
-    visibility: hidden;
-}
-.slick-slide
-{
-    display: none;
-    float: left;
-    min-height: 1px;
-}
-[dir='rtl'] .slick-slide
-{
-    float: right;
-}
-.slick-slide img
-{
-    display: block;
-}
-.slick-slide.slick-loading img
-{
-    display: none;
-}
-.slick-slide.dragging img
-{
-    pointer-events: none;
-}
-.slick-initialized .slick-slide
-{
-    display: block;
-}
-.slick-loading .slick-slide
-{
-    visibility: hidden;
-}
-.slick-vertical .slick-slide
-{
-    display: block;
-    height: auto;
-    border: 1px solid transparent;
-}
-.slick-arrow.slick-hidden {
-    display: none;
-}
-.slick-prev,
-.slick-next {
-    display: block;
-    height: 40px;
-    width: 80px;
-    background: transparent;
-    position: absolute;
-    top: 0;
-    cursor: pointer;
-    text-indent: 100px;
-    overflow: hidden;
-    border: 0;
-    outline: 0;
-    z-index: 100;
-    background-image: linear-gradient(to left, rgba(0, 0, 0, 0) 0%, rgba(20, 20, 20, 0.4) 13px, #141414 27px);
-}
-.slick-prev::after,
-.slick-next::after {
-    font: normal normal normal 18px/1 FontAwesome;
-    position: absolute;
-    top: 13px;
-    color: #d3d3d3;
-    text-indent: 0;
-}
-.slick-prev::before,
-.slick-next::before {
-    content: "";
-    height: 100%;
-    width: 1px;
-    background-image: linear-gradient(rgba(63, 63, 63, 0) 0%, rgba(63, 63, 63, 0.7) 16px, rgba(63, 63, 63, 0.7) 37px, rgba(63, 63, 63, 0) 100%);
-    position: absolute;
-    top: 0px;
-}
-.slick-prev::before {
-    left: 40px;
-}
-.slick-next::before {
-    right: 35px;
-}
-.slick-prev.slick-disabled,
-.slick-next.slick-disabled {
-    display: none !important;
-}
-.slick-prev {
-    left: 0px;
-}
-.slick-next {
-    right: 0px;
-    background-image: linear-gradient(to right, rgba(0, 0, 0, 0) 0%, rgba(20, 20, 20, 0.4) 13px, #141414 27px);
-}
-.slick-next svg,
-.slick-prev svg {
-    position: absolute;
-    height: 40px;
-    width: 1em;
-}
-.slick-next svg {
-    right: 0;
-}
-.slick-prev svg {
-    left: 0;
-}
