--- packages/neos-ui/src/Containers/EditModePanel/style.css
@@ -1,44 +1,0 @@
-.editModePanel {
-    position: fixed;
-    top: 41px; /*-70px;*/
-    left: 0;
-    padding: 0 370px; /*leave some space for slider button --size-SidebarWidth + 50px*/
-    width: 100%;
-    height: 110px;
-    background: var(--colors-ContrastDarkest);
-    transition: var(--transition-Default) ease-in-out;
-    will-change: padding;
-    transform: translateX(0);
-    opacity: 1;
-}
-.editModePanel__wrapper {
-    display: flex;
-    width: 100%;
-    padding-left: 2rem;
-}
-.editModePanel--isFringeLeft {
-    padding-left: 46px;
-}
-.editModePanel--isFringeRight {
-    padding-right: var(--spacing-GoldenUnit);
-}
-.editModePanel--isHidden {
-    display: none;
-    transform: translateX(-100%);
-    opacity: 0;
-}
-
-.editModePanel__previewModes,
-.editModePanel__editingModes {
-    width: 50%;
-    margin-right: 2rem;
-}
-
-.editModePanel__buttonWrapper {
-    margin-right: 1rem;
-    white-space: nowrap;
-}
-
-.editModePanel__current {
-    color: var(--colors-PrimaryBlue);
-}
