--- packages/neos-ui/src/Containers/PrimaryToolbar/EditModePanelToggler/index.spec.js
@@ -1,3 +1,0 @@
-test(`write tests`, () => {
-    expect(true).toBe(true);
-});
