--- packages/neos-ui/src/Containers/PrimaryToolbar/EditModePanelToggler/style.css
@@ -1,20 +1,0 @@
-.btn {
-    vertical-align: top;
-    border-right: 1px solid var(--colors-ContrastDark);
-    border-left: 1px solid var(--colors-ContrastDark);
-
-    &:focus {
-        outline: none;
-    }
-
-    @media(max-width: 768px) {
-        display: none;
-    }
-}
-
-.btn--isActive {
-    transform: translateY(1px);
-    padding-bottom: 2px;
-    color: var(--colors-PrimaryBlue);
-    background-color: var(--colors-ContrastDarkest);
-}
